Send negative response when the sub-function is missing
Services that require a sub-function do not currently return a
negative response when the sub-function is missing from the request.
This fix removes the length check from evaluateServiceResponse() and
relies on the fact that the length checks are implemented correctly
in the service callbacks.

Also re-ordered the switch statement to make it clearer which services
aren't implemented. In reality these services will never be checked in
the switch stattement as getServiceForSID() will return NULL and
trigger an early exit.

Signed-off-by: Mark Corbin <mcorbin@lunarenergy.com>